requirements.txt
fastapi>=0.115.0
uvicorn>=0.31.1
langgraph>=0.2.60
langchain-google-genai>=2.0.8
shapely>=2.0.6
jsonschema>=4.23.0
python-dotenv>=1.0.1
pydantic>=2.11.0
